+++ src/components/Input/styles.ts
@@ -22,13 +22,15 @@ export default ({ sizes: { input }, palette }: Theme) =>
     },
     counter: {
       color: palette.grey.main,
-      fontSize: '0.625rem'
+      fontSize: '0.625em',
+      lineHeight: '1em'
     },
     counterNegative: {
       color: palette.red.main
     },
     counterMultiline: {
       alignSelf: 'flex-end',
-      marginBottom: '0.3125rem'
+      height: '0.625em',
+      marginBottom: '-0.25em'
     }
   })