+++ src/components/Modal/story/index.jsx
@@ -21,15 +21,66 @@ page
 
 page
   .createChapter()
-  .addExample('Modal/story/Default.example.jsx', {
-    title: 'Default',
-    description: `
+  .addExample(
+    'Modal/story/Default.example.jsx',
+    {
+      title: 'Default',
+      description: `
 To show the 'Modal' component you should use 'useModals' hook. And if you need
 any additional logic inside the 'Modal' component you should create a wrapper 
 component and manage the internal state there.
 `
-  }) // picasso-skip-visuals
-  .addExample('Modal/story/Sizes.example.jsx', 'Sizes') // picasso-skip-visuals
-  .addExample('Modal/story/MaxHeight.example.jsx', {
-    title: 'Max Height'
-  }) // picasso-skip-visuals
+    },
+    {
+      effect: async (page, makeScreenshot) => {
+        await page.click('[data-testid="open"]')
+        await page.waitFor(100)
+        await makeScreenshot({
+          isFullScreen: true
+        })
+
+        await page.click('[data-testid="close"]')
+        await page.waitFor(100)
+        await makeScreenshot({
+          isFullScreen: true
+        })
+
+        await page.waitFor(1000)
+        await makeScreenshot()
+      }
+    }
+  )
+  .addExample('Modal/story/Sizes.example.jsx', 'Sizes', {
+    effect: async (page, makeScreenshot) => {
+      await page.click('[data-testid="trigger-small"]')
+      await page.waitFor('[data-testid="cancel"]')
+      await makeScreenshot({
+        isFullScreen: true
+      })
+
+      await page.click('[data-testid="cancel"]')
+
+      await page.click('[data-testid="trigger-medium"]')
+      await page.waitFor('[data-testid="cancel"]')
+      await makeScreenshot({
+        isFullScreen: true
+      })
+
+      await page.click('[data-testid="cancel"]')
+
+      await page.click('[data-testid="trigger-large"]')
+      await page.waitFor('[data-testid="cancel"]')
+      await makeScreenshot({
+        isFullScreen: true
+      })
+    }
+  })
+  .addExample('Modal/story/MaxHeight.example.jsx', 'Max Height', {
+    effect: async (page, makeScreenshot) => {
+      await page.click('[data-testid="trigger"]')
+      await page.waitFor('[data-testid="cancel"]')
+      await makeScreenshot({
+        isFullScreen: true
+      })
+    }
+  })